Fix json AST output
Closes #76
