+++ CommonMark/__init__.py
@@ -1,6 +1,7 @@
 # flake8: noqa
 from __future__ import unicode_literals, absolute_import
 
-from CommonMark.main import commonmark, dumpAST, ASTtoJSON
+from CommonMark.main import commonmark
+from CommonMark.dump import dumpAST, dumpJSON
 from CommonMark.blocks import Parser
 from CommonMark.render.html import HtmlRenderer
